+++ src/main/java/com/fps/svmes/models/sql/taskSchedule/DispatchRawMaterial.java
@@ -14,7 +14,7 @@ public class DispatchRawMaterial {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "dispatch_id", nullable = false)
@@ -25,7 +25,7 @@ public class DispatchRawMaterial {
     private RawMaterial rawMaterial;
 
     @Column(name = "status", nullable = false, columnDefinition = "SMALLINT DEFAULT 1")
-    private Integer status = 1; // Default active
+    private Short status = 1; // Default active
 
     public DispatchRawMaterial(Dispatch dispatch, RawMaterial rawMaterial) {
         this.dispatch = dispatch;
